Move subcommands from internal/cmd to cmd/ww.
