+++ cmd/ww/main.go
@@ -20,9 +20,9 @@ import (
 	"golang.org/x/exp/slog"
 
 	ww "github.com/wetware/pkg"
-	"github.com/wetware/pkg/internal/ls"
-	"github.com/wetware/pkg/internal/run"
-	"github.com/wetware/pkg/internal/start"
+	"github.com/wetware/pkg/cmd/ww/ls"
+	"github.com/wetware/pkg/cmd/ww/run"
+	"github.com/wetware/pkg/cmd/ww/start"
 )
 
 var flags = []cli.Flag{
